+++ posturight/lib/home.dart
@@ -6,6 +6,7 @@ import 'package:calender_picker/calender_picker.dart';
 import 'package:simple_circular_progress_bar/simple_circular_progress_bar.dart';
 import 'colors.dart';
 import 'main.dart';
+import 'title.dart';
 
 class HomePage extends StatefulWidget {
   const HomePage({required this.title, Key? key, Function? refresh}) : super(key: key);
@@ -99,7 +100,10 @@ class _HomePageState extends State<HomePage> {
               onPressed: (){
                 // Navigator.push(context,
                 //   MaterialPageRoute(builder: (context) => LoginScreen()));
-                
+                Navigator.pushAndRemoveUntil(context,
+                            MaterialPageRoute(builder: (context) => TitleScreen()),
+                            (Route<dynamic> route) => false,
+                          );
               }, 
               child: Text("Logout"),
             ),
